chore: green main CI and split FlowGraphViewInner
Fix invalid download-artifact SHAs, use docker compose on Ubuntu runners, exclude MCP unit tests from main push sweep, and harden MCP DB fixtures. Extract FlowGraph and traceability matrix modules to satisfy LOC guards.
