+++ .github/workflows/contract-tests.yml
@@ -92,7 +92,7 @@ jobs:
         go-version: '1.24'
         cache-dependency-path: backend/go.sum
     - name: Download pact files
-      uses: actions/download-artifact@d7aade707a79a3bc4130f2605987cbbf2ecfac6c
+      uses: actions/download-artifact@v4
       with:
         name: pacts
         path: tests/contracts/pacts
@@ -137,7 +137,7 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683
     - name: Download pact files
-      uses: actions/download-artifact@d7aade707a79a3bc4130f2605987cbbf2ecfac6c
+      uses: actions/download-artifact@v4
       with:
         name: pacts
         path: tests/contracts/pacts
@@ -185,7 +185,7 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683
     - name: Download pact files
-      uses: actions/download-artifact@d7aade707a79a3bc4130f2605987cbbf2ecfac6c
+      uses: actions/download-artifact@v4
       with:
         name: pacts
         path: tests/contracts/pacts
