+++ .github/workflows/openapi-docs.yml
@@ -142,7 +142,7 @@ jobs:
         persist-credentials: true
         token: ${{ secrets.GITHUB_TOKEN }}
     - name: Download OpenAPI spec
-      uses: actions/download-artifact@d7aade707a79a3bc4130f2605987cbbf2ecfac6c
+      uses: actions/download-artifact@v4
       with:
         name: openapi-spec
         path: .
@@ -177,7 +177,7 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683
     - name: Download OpenAPI spec
-      uses: actions/download-artifact@d7aade707a79a3bc4130f2605987cbbf2ecfac6c
+      uses: actions/download-artifact@v4
       with:
         name: openapi-spec
         path: .
