+++ .github/workflows/test-validation.yml
@@ -204,7 +204,7 @@ jobs:
     steps:
     - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683
     - name: Download all artifacts
-      uses: actions/download-artifact@d7aade707a79a3bc4130f2605987cbbf2ecfac6c
+      uses: actions/download-artifact@v4
       with:
         path: validation-artifacts/
     - name: Generate consolidated report
