+++ .github/workflows/test.yml
@@ -176,11 +176,9 @@ jobs:
     - name: Set up Docker Buildx
       uses: docker/setup-buildx-action@8d2750c68a42422c14e847fe6c8ac0403b4cbd6f
     - name: Start services
-      run: 'docker-compose up -d
-
+      run: |
+        docker compose up -d
         sleep 10
-
-        '
     - name: Run integration tests
       run: '# Add integration test commands here
 
@@ -189,5 +187,5 @@ jobs:
         '
     - name: Stop services
       if: always()
-      run: docker-compose down
+      run: docker compose down
     timeout-minutes: 30
