+++ .github/workflows/tests.yml
@@ -107,7 +107,9 @@ jobs:
           if [ "${{ github.event_name }}" = "pull_request" ]; then
             mapfile -t unit_targets < <(bash .github/scripts/pr_unit_test_targets.sh "${{ github.event.pull_request.base.sha }}")
           else
-            unit_targets=(tests/unit)
+            mapfile -t unit_targets < <(
+              find tests/unit -type f -name 'test_*.py' ! -path 'tests/unit/mcp/*' | sort
+            )
           fi
           if ((${#unit_targets[@]} == 0)); then
             unit_targets=(
@@ -116,11 +118,11 @@ jobs:
             )
           fi
           echo "Unit pytest targets: ${unit_targets[*]}" >> "$GITHUB_STEP_SUMMARY"
-          uv run pytest "${unit_targets[@]}" --ignore=src/tracertm/mcp -m "not slow" -v --cov=src/tracertm --cov-report=xml \
+          uv run pytest "${unit_targets[@]}" --ignore=tests/unit/mcp --ignore=src/tracertm/mcp -m "not slow" -v --cov=src/tracertm --cov-report=xml \
             --cov-report=html --cov-report=term-missing --cov-fail-under=0 \
             --junitxml=junit-${{ matrix.test-type }}.xml
           set +e
-          uv run pytest "${unit_targets[@]}" --ignore=src/tracertm/mcp -m "slow" -v --cov=src/tracertm --cov-append \
+          uv run pytest "${unit_targets[@]}" --ignore=tests/unit/mcp --ignore=src/tracertm/mcp -m "slow" -v --cov=src/tracertm --cov-append \
             --cov-report=xml --cov-report=html --cov-report=term-missing --cov-fail-under=0 \
             --junitxml=junit-${{ matrix.test-type }}-slow.xml
           slow_code=$?
