+++ tests/unit/mcp/conftest.py
@@ -65,6 +65,7 @@ async def setup_test_database(
     engine = await get_async_engine()
 
     async with engine.begin() as conn:
+        await conn.run_sync(Base.metadata.drop_all)
         await conn.run_sync(Base.metadata.create_all)
 
     yield
